refactor(editor): move selection away from flexible view into its own component
