+++ components/calendar-editor-week-view/calendarEditorWeek.tsx
@@ -15,6 +15,7 @@ import UpdateEventForm from "@/components/event-update-form/updateEvent";
 import { Actions } from "@/hooks/mapHook";
 import { FlexibleView, ViewSize } from "../shared/flexible-view/flexibleView";
 import { DayViewContent } from "../shared/day-view/dayContent";
+import { Selection } from "../editor-selection-view/editorSelection";
 
 const CALENDAR_WEEK_CONTAINER_ID = "calendar-week-container";
 
@@ -185,6 +186,7 @@ const CalendarEditorWeek = ({
             )}
           />
         </div>
+        <Selection />
       </FlexibleView>
       {selectedEvent.mapOrElse(
         () => null,
